feat: add update interface
